// frontend/src/services/model/userWithStats.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Football Pool API
 * OpenAPI spec version: 1.0.0
 */
import type { PlayerResponse } from "./playerResponse";

export interface UserWithStats {
  id: number;
  name: string;
  email: string;
  role: string;
  player?: PlayerResponse;
  pick_count: number;
  total_wins: number;
  created_at: string;
  updated_at: string;
}